兼容 oracle 改造

+++ api/src/main/java/top/cardone/authority/po/Role.java
@@ -120,7 +120,7 @@ public class Role implements java.io.Serializable {
      */
     @lombok.Getter
     @lombok.Setter
-    protected Long orderBy;
+    protected Integer orderBy;
 
     /**
      * 组织编号
