+++ api/src/main/java/top/cardone/authority/po/UserGroupUser.java
@@ -113,7 +113,7 @@ public class UserGroupUser implements java.io.Serializable {
      */
     @lombok.Getter
     @lombok.Setter
-    protected Long orderBy;
+    protected Integer orderBy;
 
     /**
      * 组织编号
